requirements.txt
pydantic
python-dotenv
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
defusedxml
pypdf
faiss-cpu
langchain-google-genai
pinecone-client
wikipedia
bs4
requests
streamlit
langchain-huggingface
langchain-pinecone
pdfplumber
langgraph
sentence-transformers